Modernize other areas of the UI
